+++ packages/cli/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@celo/celocli",
   "description": "CLI Tool for transacting with the Celo protocol",
-  "version": "0.0.48-dev",
+  "version": "0.0.48-sami",
   "author": "Celo",
   "license": "Apache-2.0",
   "repository": "celo-org/celo-monorepo",
@@ -31,8 +31,8 @@
     "test": "TZ=UTC jest --runInBand"
   },
   "dependencies": {
-    "@celo/contractkit": "0.4.3-dev",
-    "@celo/utils": "0.1.14-dev",
+    "@celo/contractkit": "https://github.com/mrsmkl/celo-monorepo#celo-contractkit-v0.4.3-sami-gitpkg",
+    "@celo/utils": "0.1.13",
     "@ledgerhq/hw-transport-node-hid": "^5.11.0",
     "@oclif/command": "^1",
     "@oclif/config": "^1",
