+++ packages/contractkit/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@celo/contractkit",
-  "version": "0.4.3-dev",
+  "version": "0.4.3-sami",
   "description": "Celo's ContractKit to interact with Celo network",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
@@ -30,7 +30,7 @@
     "@azure/identity": "^1.0.2",
     "@azure/keyvault-keys": "^4.0.2",
     "@azure/keyvault-secrets": "^4.0.2",
-    "@celo/utils": "0.1.14-dev",
+    "@celo/utils": "0.1.13",
     "@ledgerhq/hw-app-eth": "^5.11.0",
     "@ledgerhq/hw-transport": "^5.11.0",
     "@ledgerhq/hw-transport-node-hid": "^5.11.0",
